+++ lib/suspenders/app_builder.rb
@@ -278,8 +278,8 @@ Rack::Timeout.timeout = (ENV["RACK_TIMEOUT"] || 10).to_i
       generate 'rspec:install'
     end
 
-    def configure_unicorn
-      copy_file 'unicorn.rb', 'config/unicorn.rb'
+    def configure_puma
+      copy_file "puma.rb", "config/puma.rb"
     end
 
     def setup_foreman