+++ lib/suspenders/generators/app_generator.rb
@@ -138,7 +138,7 @@ module Suspenders
       build :disable_xml_params
       build :fix_i18n_deprecation_warning
       build :setup_default_rake_task
-      build :configure_unicorn
+      build :configure_puma
       build :setup_foreman
     end